Stage

+++ services/meeting-socket-service/src/modules/meetings/meetings.common.ts
@@ -16,6 +16,7 @@ import { MeetingDocument } from '../../schemas/meeting.schema';
 import { MeetingUserDocument } from '../../schemas/meeting-user.schema';
 import { MeetingChatsService } from '../meeting-chats/meeting-chats.service';
 import { MeetingChatReactionsService } from '../meeting-chats/meeting-chat-reactions.service';
+import { MeetingReactionsService } from '../meeting-reactions/meeting-reactions.service';
 import { MeetingQuestionAnswersService } from '../meeting-question-answer/meeting-question-answer.service';
 import { MeetingQuestionAnswerReactionsService } from '../meeting-question-answer/meeting-question-answer-reactions.service';
 
@@ -27,6 +28,7 @@ export class MeetingsCommonService {
     private coreService: CoreService,
     private usersService: UsersService,
     private readonly meetingChatsService: MeetingChatsService,
+    private readonly meetingReactionsService: MeetingReactionsService,
     private readonly meetingChatReactionsService: MeetingChatReactionsService,
     private readonly meetingQuestionAnswersService: MeetingQuestionAnswersService,
     private readonly meeetingQuestionAnswerReactionsService: MeetingQuestionAnswerReactionsService,
@@ -89,6 +91,11 @@ export class MeetingsCommonService {
       session,
     });
 
+    await this.meetingReactionsService.deleteMany({
+      query: { meeting: meetingId },
+      session,
+    });
+
     await this.meeetingQuestionAnswerReactionsService.deleteMany({
       query: { meeting: meetingId },
       session,
@@ -103,9 +110,9 @@ export class MeetingsCommonService {
       query: { meeting: meetingId },
       session,
     });
-    await this.usersService.deleteMany({ meeting: meetingId }, session);
+    await this.usersService.deleteMany({ meeting: meetingId, accessStatus: { $ne: MeetingAccessStatusEnum.Waiting } }, session);
 
-    await this.meetingsService.deleteById({ meetingId }, session);
+    // await this.meetingsService.deleteById({ meetingId }, session);
   }
 
   async handleClearMeetingData({
@@ -116,12 +123,12 @@ export class MeetingsCommonService {
     session,
   }) {
     await this.clearMeeting({ meetingId, session });
-    await this.coreService.updateMeetingInstance({
-      instanceId,
-      data: {
-        owner: null,
-      },
-    });
+    // await this.coreService.updateMeetingInstance({
+    //   instanceId,
+    //   data: {
+    //     owner: null,
+    //   },
+    // });
     await this.coreService.updateUserTemplate({
       templateId,
       userId,
